@@ -47,13 +47,15 @@
   },
   "devDependencies": {
     "@babel/core": "^7.5.5",
+    "babel-jest": "^24.9.0",
     "babel-loader": "^8.0.6",
+    "babel-plugin-transform-dynamic-import": "^2.1.0",
     "babel-plugin-transform-es3-member-expression-literals": "^6.22.0",
     "babel-plugin-transform-es3-property-literals": "^6.22.0",
     "babel-preset-env": "^1.7.0",
     "clean-webpack-plugin": "^3.0.0",
     "copy-webpack-plugin": "^5.0.4",
-    "eslint": "^6.1.0",
+    "eslint": "^6.2.1",
     "eslint-plugin-jest": "^22.13.7",
     "gh-pages": "^2.0.1",
     "gitbook-cli": "^2.3.2",
